the player is alive!!!
CH 3.3.5 builds < 287 do not properly check offline player names when the player is using a fake UUID (i.e. Bedrock via geyser/floodgate), returning player not found, which was throwing the account functions in this script - moved to using UUID.
This bug was fixed in CH build #287, so latest builds can use player names in this context once more.