bot: cmd: add Whitelist
Whitelists a player, but only if the discord user has either admin roles or an approved role (eg set to only those with Minecraft role, etc).
If not an admin and uses exceed configured limits the whitelist wont go through either.
logs various infos to storage, eg log of whitelists and who initiated.
Note this is just the vanilla whitelist system and probably doesn't support a plugin's whitelist.