Bot: Send Minecraft player Chat to Multiple guilds
Bot now sends minecraft player's chat to all servers, if messages are enabled.

Adds permission check for channels, cannot send or receive on a channel without permission (unless default)

Note: If your server is not set as the primary guild the channel override and using channels other than the minecraft channel are ignored.

Fixes:
- Roles rank was checked on every message, now only checks once per day.
- Player events moved to new file, more events notifications to come.